update build instructions

+++ Tools/Src/BuildAll.cs
@@ -10,7 +10,7 @@ public static partial class Funcs
         if (bootRc != 0) return bootRc;
 
         var buildBinDir = Path.GetFullPath(Path.Combine(G.repoRoot, G.cfg.BuildDir, "bin", G.buildConfig.ToString()));
-        if (Directory.Exists(buildBinDir))
+        if (!G.skipCleanBuildDir && Directory.Exists(buildBinDir))
         {
             Directory.Delete(buildBinDir, recursive: true);
             LogInfo($"BuildAll: cleaned {buildBinDir}");
